refactor: RecommendRestaurantButton을 KakaoMap 폴더로 이전
