+++ src/components/KakaoMap/index.tsx
@@ -9,9 +9,9 @@ import { kakaoMapOptionsState } from 'stores/kakaoMap';
 import { getElement } from 'utils/helpers/elementHandler';
 import { kakaoMapAddEventListener, kakaoMapHelpers } from 'utils/helpers/kakaoMap';
 
-import RecommendRandomRestaurantButton from '../common/Buttons/RecommendRandomRestaurantButton';
 import RestaurantBottomDrawer from '../Restaurant/RestaurantBottomDrawer';
 import CurrentLocationButton from './CurrentLocationButton';
+import RecommendRandomRestaurantButton from './RecommendRandomRestaurantButton';
 import ZoomInButton from './ZoomInButton';
 import ZoomOutButton from './ZoomOutButton';
